pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-company"
dynamic = ["version"]
description = "AI Agent Orchestration System with Guardrails, Memory, and Observability"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Agentic Company Team", email = "team@agentic.company"}
]
keywords = ["ai", "agents", "orchestration", "llm", "guardrails"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "tenacity>=8.2.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-timeout>=2.2.0",
    "black>=24.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "httpx>=0.26.0",
]
anthropic = [
    "anthropic>=0.18.0",
]
openai = [
    "openai>=1.10.0",
]
supabase = [
    "supabase>=2.3.0",
]
celery = [
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
]
postgres = [
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
]
observability = [
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "prometheus-client>=0.19.0",
]
diagnostics = [
    "playwright>=1.40.0",
]
test-diagnostics = [
    "pytest-playwright>=0.4.0",
]
all = [
    "agentic-company[dev,anthropic,openai,supabase,celery,postgres,observability,diagnostics]",
]

[project.scripts]
agentic = "orchestration.cli:main"
agenticom = "agenticom.cli:main"
agenticom-launch = "orchestration.launcher:main"

[project.urls]
Homepage = "https://github.com/wjlgatech/agentic-company"
Documentation = "https://agentic.company/docs"
Repository = "https://github.com/wjlgatech/agentic-company"
Issues = "https://github.com/wjlgatech/agentic-company/issues"

[tool.setuptools.dynamic]
version = {attr = "orchestration._version.__version__"}

[tool.setuptools.packages.find]
where = ["."]
include = ["orchestration*", "agenticom*"]

[tool.setuptools.package-data]
agenticom = ["bundled_workflows/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "integration: marks tests that require live API keys or external services (deselect with '-m not integration')",
]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = [
    "E501",   # Line too long — handled by Black
    "N818",   # Exception names should end in *Error — renaming is a breaking API change
    "B008",   # Function call in default arg — intentional singleton pattern in security.py
    "B904",   # raise-without-from-clause — TODO: add `from err` across codebase
    "E741",   # Ambiguous variable name (l, O, I) — TODO: rename in lessons.py
    "E721",   # Type comparison — TODO: migrate to isinstance()
    "N806",   # Uppercase variable in function — TODO: refactor FailureHandler pattern
    "F821",   # Undefined name — TODO: fix FailureHandler forward reference in agenticom/workflows.py
]

[tool.ruff.lint.per-file-ignores]
# These files are excluded from pytest collection (require real LLMs / browsers /
# hardcoded paths). Ruff still sees them, so suppress their structural violations.
"tests/test_complete_integration.py" = ["E402", "F401"]
"tests/test_stage_tracking.py" = ["E402"]
"tests/test_real_cases.py" = ["E722", "E402"]
# Lazy / conditional imports that are intentional
"orchestration/integrations/openclaw.py" = ["F401"]
"orchestration/integrations/nanobot.py" = ["F401"]
"orchestration/api.py" = ["E402"]
"agenticom/workflows.py" = ["E402"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
